fix: update status to table subject end teaching schedules, update method delete subject

+++ backend/app/services/subjects.py
@@ -13,8 +13,8 @@ from app.models.schemas.subjects.subject_schemas import (
     SubjectDeleteResponse
 )
 from app.enums.status import StatusEnum
-# from app.models.models import LearningSchedules
-# from app.models.models import ExaminationSchedules
+from app.models.models import LearningSchedules
+from app.models.models import ExaminationSchedules
 
 class SubjectServices:
     @staticmethod
@@ -94,15 +94,27 @@ class SubjectServices:
             raise HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND, detail="Subject not found"
             )
+
+        check_related_entities = select(LearningSchedules).where(LearningSchedules.subject_id == subject_id)
+        learningSchedules = session.exec(check_related_entities).all()
+        if learningSchedules:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Subject has related LearningSchedules and cannot be deleted.",
+            )
         
-        # todo: check related 
+        check_related_entities_second = select(ExaminationSchedules).where(ExaminationSchedules.subject_id == subject_id)
+        examinationSchedules = session.exec(check_related_entities_second).all()
+        if examinationSchedules:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Subject has related ExaminationSchedules and cannot be deleted.",
+            )
 
-
-        # add status to subject
-        # if subject.status == StatusEnum.ACTIVE:
-        #     subject.status = StatusEnum.INACTIVE
-        #     session.commit()
-        #     return SubjectDeleteResponse(id=str(subject.id), message="subject set to inactive")
+        if subject.status == StatusEnum.ACTIVE:
+            subject.status = StatusEnum.INACTIVE
+            session.commit()
+            return SubjectDeleteResponse(id=str(subject.id), message="subject set to inactive")
         
         session.delete(subject)
         session.commit()
